refactor(button-group): rename FilterButtonGroup to ButtonGroup and move to src
