+++ packages/components/src/Filter/components/FilterButtonRemovable/FilterButtonRemovable.tsx
@@ -2,6 +2,7 @@ import React, { forwardRef } from "react"
 import { Icon } from "@kaizen/component-library"
 import iconClear from "@kaizen/component-library/icons/clear.icon.svg"
 import { Tooltip } from "@kaizen/draft-tooltip"
+import { ButtonGroup, ButtonGroupProps } from "../../../ButtonGroup"
 import { DataAttributes } from "../../../types"
 import { isRefObject } from "../../../utils/isRefObject"
 import {
@@ -9,13 +10,9 @@ import {
   FilterTriggerButtonProps,
 } from "../FilterTriggerButton"
 import { FilterButton } from "../_primitives/FilterButton"
-import {
-  FilterButtonGroup,
-  FilterButtonGroupProps,
-} from "../_primitives/FilterButtonGroup"
 
 export interface FilterButtonRemovableProps
-  extends Omit<FilterButtonGroupProps, "children"> {
+  extends Omit<ButtonGroupProps, "children"> {
   triggerButtonProps: FilterTriggerButtonProps & DataAttributes
   removeButtonProps: Partial<Omit<FilterTriggerButtonProps, "children">> &
     DataAttributes & { tooltipText?: string }
@@ -38,14 +35,14 @@ export const FilterButtonRemovable = forwardRef<
     `Remove filter - ${triggerButtonProps?.label}`
 
   return (
-    <FilterButtonGroup {...restProps}>
+    <ButtonGroup {...restProps}>
       <FilterTriggerButton ref={ref} {...triggerButtonProps} />
       <Tooltip text={removeButtonLabel} display="inline-block" position="below">
         <FilterButton ref={removeButtonRef} {...removeButtonProps}>
           <Icon icon={iconClear} title={removeButtonLabel} />
         </FilterButton>
       </Tooltip>
-    </FilterButtonGroup>
+    </ButtonGroup>
   )
 })
