--- packages/components/src/Filter/components/_primitives/FilterButtonGroup/index.ts
@@ -1,1 +1,0 @@
-export * from "./FilterButtonGroup"
